refactor: move SpreadsheetEdits
